chore: fix testing

+++ src/product/dto/create-product.input.ts
@@ -1,4 +1,4 @@
-import { Field, InputType, Int } from "@nestjs/graphql";
+import { Field, Float, InputType } from "@nestjs/graphql";
 import { IsNotEmpty, IsNumber, IsString } from "class-validator";
 import { ICreateProduct } from "../product.interface";
 
@@ -16,7 +16,7 @@ export class CreateProductInput implements ICreateProduct {
 
   @IsNotEmpty()
   @IsNumber()
-  @Field(() => Int)
+  @Field(() => Float)
   price!: number;
 
   @IsNotEmpty()
