+++ src/product/dto/product.object.ts
@@ -1,4 +1,4 @@
-import { Field, ID, Int, ObjectType } from "@nestjs/graphql";
+import { Field, Float, ID, ObjectType } from "@nestjs/graphql";
 import { IProduct } from "../product.interface";
 
 @ObjectType()
@@ -12,7 +12,7 @@ export class Product implements IProduct {
   @Field()
   description!: string;
 
-  @Field(() => Int)
+  @Field(() => Float)
   price!: number;
 
   @Field()
